+++ src/modules/olMap/tooltip/HelpTooltip.js
@@ -1,7 +1,7 @@
 
-import { BaOverlay, BaOverlayTypes } from './components/BaOverlay';
+import { BaOverlay, BaOverlayTypes } from '../components/BaOverlay';
 import Overlay from 'ol/Overlay';
-import { $injector } from '../../injection';
+import { $injector } from '../../../injection';
 
 export class HelpTooltip {
